+++ mythtv/bindings/python/MythTV/ttvdb/tvdb_api.py
@@ -15,13 +15,6 @@ import warnings
 import logging
 import datetime
 
-from .tvdb_ui import BaseUI, ConsoleUI
-from .tvdb_exceptions import (
-    tvdb_error, tvdb_notauthorized, tvdb_shownotfound,
-    tvdb_seasonnotfound, tvdb_episodenotfound,
-    tvdb_resourcenotfound, tvdb_invalidlanguage,
-    tvdb_attributenotfound
-)
 
 """Simple-to-use Python interface to The TVDB's API (thetvdb.com)
 
@@ -33,15 +26,18 @@ Example usage:
 u'Cabin Fever'
 """
 __author__ = "dbr/Ben"
-__version__ = "1.11"
+__version__ = "2.0-dev"
 
 
 IS_PY2 = sys.version_info[0] == 2
 
 if IS_PY2:
+    user_input = raw_input
     from urllib import quote as url_quote
 else:
     from urllib.parse import quote as url_quote
+    user_input = input
+
 
 if IS_PY2:
     int_types = (int, long)
@@ -56,6 +52,210 @@ lastTimeout = None
 def log():
     return logging.getLogger("tvdb_api")
 
+
+## Exceptions
+
+class tvdb_exception(Exception):
+    """Any exception generated by tvdb_api
+    """
+    pass
+
+class tvdb_error(tvdb_exception):
+    """An error with thetvdb.com (Cannot connect, for example)
+    """
+    pass
+
+class tvdb_userabort(tvdb_exception):
+    """User aborted the interactive selection (via
+    the q command, ^c etc)
+    """
+    pass
+
+class tvdb_notauthorized(tvdb_exception):
+    """An authorization error with thetvdb.com
+    """
+    pass
+
+class tvdb_shownotfound(tvdb_exception):
+    """Show cannot be found on thetvdb.com (non-existant show)
+    """
+    pass
+
+class tvdb_seasonnotfound(tvdb_exception):
+    """Season cannot be found on thetvdb.com
+    """
+    pass
+
+class tvdb_episodenotfound(tvdb_exception):
+    """Episode cannot be found on thetvdb.com
+    """
+    pass
+
+class tvdb_resourcenotfound(tvdb_exception):
+    """Resource cannot be found on thetvdb.com
+    """
+    pass
+
+class tvdb_invalidlanguage(tvdb_exception):
+    """invalid language given on thetvdb.com
+    """
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)
+
+class tvdb_attributenotfound(tvdb_exception):
+    """Raised if an episode does not have the requested
+    attribute (such as a episode name)
+    """
+    pass
+
+
+## UI
+
+class BaseUI(object):
+    """Base user interface for Tvdb show selection.
+
+    Selects first show.
+
+    A UI is a callback. A class, it's __init__ function takes two arguments:
+
+    - config, which is the Tvdb config dict, setup in tvdb_api.py
+    - log, which is Tvdb's logger instance (which uses the logging module). You can
+    call log.info() log.warning() etc
+
+    It must have a method "selectSeries", this is passed a list of dicts, each dict
+    contains the the keys "name" (human readable show name), and "sid" (the shows
+    ID as on thetvdb.com). For example:
+
+    [{'name': u'Lost', 'sid': u'73739'},
+     {'name': u'Lost Universe', 'sid': u'73181'}]
+
+    The "selectSeries" method must return the appropriate dict, or it can raise
+    tvdb_userabort (if the selection is aborted), tvdb_shownotfound (if the show
+    cannot be found).
+
+    A simple example callback, which returns a random series:
+
+    >>> import random
+    >>> from tvdb_ui import BaseUI
+    >>> class RandomUI(BaseUI):
+    ...    def selectSeries(self, allSeries):
+    ...            import random
+    ...            return random.choice(allSeries)
+
+    Then to use it..
+
+    >>> from tvdb_api import Tvdb
+    >>> t = Tvdb(custom_ui = RandomUI)
+    >>> random_matching_series = t['Lost']
+    >>> type(random_matching_series)
+    <class 'tvdb_api.Show'>
+    """
+    def __init__(self, config, log = None):
+        self.config = config
+        if log is not None:
+            warnings.warn("the UI's log parameter is deprecated, instead use\n"
+                "use import logging; logging.getLogger('ui').info('blah')\n"
+                "The self.log attribute will be removed in the next version")
+            self.log = logging.getLogger(__name__)
+
+    def selectSeries(self, allSeries):
+        return allSeries[0]
+
+
+class ConsoleUI(BaseUI):
+    """Interactively allows the user to select a show from a console based UI
+    """
+
+    def _displaySeries(self, allSeries, limit = 6):
+        """Helper function, lists series with corresponding ID
+        """
+        if limit is not None:
+            toshow = allSeries[:limit]
+        else:
+            toshow = allSeries
+
+        print("TVDB Search Results:")
+        for i, cshow in enumerate(toshow):
+            i_show = i + 1 # Start at more human readable number 1 (not 0)
+            log().debug('Showing allSeries[%s], series %s)' % (i_show, allSeries[i]['seriesName']))
+            if i == 0:
+                extra = " (default)"
+            else:
+                extra = ""
+
+            lid_map = dict((v, k) for (k, v) in self.config['langabbv_to_id'].items())
+
+            output = "%s -> %s [%s] # http://thetvdb.com/?tab=series&id=%s&lid=%s%s" % (
+                i_show,
+                cshow['seriesName'],
+                lid_map[cshow['lid']],
+                str(cshow['id']),
+                cshow['lid'],
+                extra
+            )
+            if IS_PY2:
+                print(output.encode("UTF-8", "ignore"))
+            else:
+                print(output)
+
+    def selectSeries(self, allSeries):
+        self._displaySeries(allSeries)
+
+        if len(allSeries) == 1:
+            # Single result, return it!
+            print("Automatically selecting only result")
+            return allSeries[0]
+
+        if self.config['select_first'] is True:
+            print("Automatically returning first search result")
+            return allSeries[0]
+
+        while True: # return breaks this loop
+            try:
+                print("Enter choice (first number, return for default, 'all', ? for help):")
+                ans = user_input()
+            except KeyboardInterrupt:
+                raise tvdb_userabort("User aborted (^c keyboard interupt)")
+            except EOFError:
+                raise tvdb_userabort("User aborted (EOF received)")
+
+            log().debug('Got choice of: %s' % (ans))
+            try:
+                selected_id = int(ans) - 1 # The human entered 1 as first result, not zero
+            except ValueError: # Input was not number
+                if len(ans.strip()) == 0:
+                    # Default option
+                    log().debug('Default option, returning first series')
+                    return allSeries[0]
+                if ans == "q":
+                    log().debug('Got quit command (q)')
+                    raise tvdb_userabort("User aborted ('q' quit command)")
+                elif ans == "?":
+                    print("## Help")
+                    print("# Enter the number that corresponds to the correct show.")
+                    print("# a - display all results")
+                    print("# all - display all results")
+                    print("# ? - this help")
+                    print("# q - abort tvnamer")
+                    print("# Press return with no input to select first result")
+                elif ans.lower() in ["a", "all"]:
+                    self._displaySeries(allSeries, limit = None)
+                else:
+                    log().debug('Unknown keypress %s' % (ans))
+            else:
+                log().debug('Trying to return ID: %d' % (selected_id))
+                try:
+                    return allSeries[selected_id]
+                except IndexError:
+                    log().debug('Invalid show number entered!')
+                    print("Invalid number (%s) selected!")
+                    self._displaySeries(allSeries)
+
+
+## Main API
 
 class ShowContainer(dict):
     """Simple dict that holds a series of Show instances
@@ -93,6 +293,16 @@ class Show(dict):
         )
 
     def __getitem__(self, key):
+        v1_compatibility = {
+            'seriesname': 'seriesName',
+        }
+
+        if key in v1_compatibility:
+            import warnings
+            msg = "v1 usage is deprecated, please use new names: old: '%s', new: '%s'" % (
+                key, v1_compatibility[key])
+            key = v1_compatibility[key]
+
         if key in self:
             # Key is an episode, return it
             return dict.__getitem__(self, key)
@@ -242,12 +452,14 @@ class Episode(dict):
             v1_compatibility = {
                 'episodenumber': 'airedEpisodeNumber',
                 'firstaired': 'firstAired',
-                'seasonnumber': 'airedSeason'
+                'seasonnumber': 'airedSeason',
+                'episodename': 'episodeName',
             }
             if key in v1_compatibility:
                 import warnings
-                msg = "v1 usage is deprecated, please use new names: old: '%s', new:'%s'"
-                warnings.warn(msg %(key, v1_compatibility[key]), category=DeprecationWarning)
+                msg = "v1 usage is deprecated, please use new names: old: '%s', new: '%s'" % (
+                    key, v1_compatibility[key])
+                warnings.warn(msg, category=DeprecationWarning)
                 try:
                     value = dict.__getitem__(self, v1_compatibility[key])
                     if key in ['episodenumber', 'seasonnumber']:
@@ -255,10 +467,11 @@ class Episode(dict):
                         return str(value)
                     else:
                         return value
-                except:
+                except KeyError:
                     # We either return something or we get the exception below
                     pass
-            raise tvdb_attributenotfound("Cannot find attribute %s" % (repr(key)))
+
+                    raise tvdb_attributenotfound("Cannot find attribute %s" % (repr(key)))
 
     def search(self, term=None, key=None):
         """Search episode data for term, if it matches, return the Episode (self).
@@ -373,19 +586,15 @@ class Tvdb:
                  >>> import logging
                  >>> logging.basicConfig(level = logging.DEBUG)
 
-        cache (True/False/str/unicode/urllib2 opener):
-            Retrieved XML are persisted to to disc. If true, stores in
-            tvdb_api folder under your systems TEMP_DIR, if set to
-            str/unicode instance it will use this as the cache
-            location. If False, disables caching.  Can also be passed
-            an arbitrary Python object, which is used as a urllib2
-            opener, which should be created by urllib2.build_opener
+        cache (True/False/str/requests_cache.CachedSession):
 
-            In Python 3, True/False enable or disable default
-            caching. Passing string specified directory where to store
-            the "tvdb.sqlite3" cache file. Also a custom
+            Retrieved URLs can be persisted to to disc.
+
+            True/False enable or disable default caching. Passing
+            string specifies the directory where to store the
+            "tvdb.sqlite3" cache file. Alternatively a custom
             requests.Session instance can be passed (e.g maybe a
-            customised instance of requests_cache.CachedSession)
+            customised instance of `requests_cache.CachedSession`)
 
         banners (True/False):
             Retrieves the banners for a show. These are accessed
@@ -464,9 +673,9 @@ class Tvdb:
             }
         else:
             self.config['auth_payload'] = {
-                "apikey": "199FD9384384C73A",
-                "userkey": "9BBA332F483717C1",
-                "username": "pelmen"
+                "apikey": "0629B785CE550C8D",
+                "userkey": "",
+                "username": ""
             }
 
         self.config['debug_enabled'] = debug  # show debugging messages
