+++ mythtv/bindings/python/MythTV/ttvdb/tvdb_exceptions.py
@@ -9,63 +9,20 @@
 """
 
 __author__ = "dbr/Ben"
-__version__ = "1.11"
+__version__ = "2.0-dev"
 
-__all__ = ["tvdb_error", "tvdb_userabort", "tvdb_shownotfound",
-"tvdb_seasonnotfound", "tvdb_episodenotfound", "tvdb_attributenotfound"]
+import logging
 
-class tvdb_exception(Exception):
-    """Any exception generated by tvdb_api
-    """
-    pass
+__all__ = ["tvdb_error", "tvdb_userabort", "tvdb_notauthorized", "tvdb_shownotfound",
+"tvdb_seasonnotfound", "tvdb_episodenotfound", "tvdb_attributenotfound",
+"tvdb_resourcenotfound", "tvdb_invalidlanguage"]
 
-class tvdb_error(tvdb_exception):
-    """An error with thetvdb.com (Cannot connect, for example)
-    """
-    pass
+logging.getLogger(__name__).warning(
+    "tvdb_exceptions module is deprecated - use classes directly from tvdb_api instead")
 
-class tvdb_userabort(tvdb_exception):
-    """User aborted the interactive selection (via
-    the q command, ^c etc)
-    """
-    pass
-
-class tvdb_notauthorized(tvdb_exception):
-    """An authorization error with thetvdb.com
-    """
-    pass
-
-class tvdb_shownotfound(tvdb_exception):
-    """Show cannot be found on thetvdb.com (non-existant show)
-    """
-    pass
-
-class tvdb_seasonnotfound(tvdb_exception):
-    """Season cannot be found on thetvdb.com
-    """
-    pass
-
-class tvdb_episodenotfound(tvdb_exception):
-    """Episode cannot be found on thetvdb.com
-    """
-    pass
-
-class tvdb_resourcenotfound(tvdb_exception):
-    """Resource cannot be found on thetvdb.com
-    """
-    pass
-
-class tvdb_invalidlanguage(tvdb_exception):
-    """invalid language given on thetvdb.com
-    """
-    def __init__(self, value):
-        self.value = value
-
-    def __str__(self):
-        return repr(self.value)
-
-class tvdb_attributenotfound(tvdb_exception):
-    """Raised if an episode does not have the requested
-    attribute (such as a episode name)
-    """
-    pass
+from tvdb_api import (
+    tvdb_error, tvdb_userabort, tvdb_notauthorized, tvdb_shownotfound,
+    tvdb_seasonnotfound, tvdb_episodenotfound,
+    tvdb_resourcenotfound, tvdb_invalidlanguage,
+    tvdb_attributenotfound
+)
